@@ -1,22 +1,24 @@
 # frozen_string_literal: true
+
 source "https://rubygems.org"
-ruby "2.3.0"
-gem "rails", "4.2.4"
-gem "rails-api"
-gem "jwt"
+ruby "2.4.0"
+
 gem "active_model_serializers"
-gem "coveralls", require: false
-gem "simple_command"
 gem "bcrypt", "~> 3.1.7"
+gem "coveralls", require: false
+gem "jwt"
+gem "rack-cors"
+gem "rails", "~> 5.2"
+gem "simple_command"
 
 group :development, :test do
-  gem "rspec-rails", "~> 3.0"
-  gem "factory_girl_rails", "~> 4.0"
-  gem "faker"
-  gem "sqlite3"
-  gem "pry"
   gem "annotate"
+  gem "factory_bot_rails"
+  gem "faker"
+  gem "pry"
+  gem "rspec-rails", "~> 3.0"
   gem "spring"
+  gem "sqlite3"
 end
 
 group :production do
