@@ -1,5 +1,6 @@
 # frozen_string_literal: true
-class CreateBucketlists < ActiveRecord::Migration
+
+class CreateBucketlists < ActiveRecord::Migration[5.2]
   def change
     create_table :bucketlists do |t|
       t.string :title
