@@ -1,5 +1,6 @@
 # frozen_string_literal: true
-class CreateItems < ActiveRecord::Migration
+
+class CreateItems < ActiveRecord::Migration[5.2]
   def change
     create_table :items do |t|
       t.text :description
