@@ -1,5 +1,6 @@
 # frozen_string_literal: true
-class CreateUsers < ActiveRecord::Migration
+
+class CreateUsers < ActiveRecord::Migration[5.2]
   def change
     create_table :users do |t|
       t.string :username
