@@ -1,5 +1,6 @@
 # frozen_string_literal: true
-class AddAuthTokenToUsers < ActiveRecord::Migration
+
+class AddAuthTokenToUsers < ActiveRecord::Migration[5.2]
   def change
     add_column :users, :auth_token, :string
   end
