@@ -1,5 +1,6 @@
 # frozen_string_literal: true
-class ChangeColumnDescriptionInItemsToName < ActiveRecord::Migration
+
+class ChangeColumnDescriptionInItemsToName < ActiveRecord::Migration[5.2]
   def change
     rename_column :items, :description, :name
   end
