@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 # == Schema Information
 #
 # Table name: bucketlists
@@ -10,7 +11,7 @@
 #  user_id    :integer
 #
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :bucketlist do
     name Faker::Hipster.word
     user_id 1
