@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 # == Schema Information
 #
 # Table name: users
@@ -11,7 +12,7 @@
 #  auth_token      :string
 #
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :user do
     username Faker::Name.first_name
     password Faker::Name.name
