@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 class AuthenticateUser
   prepend SimpleCommand
 
@@ -24,7 +25,7 @@ class AuthenticateUser
 
   def user
     user = User.find_by(username: username)
-    return user if user && user.authenticate(password)
+    return user if user&.authenticate(password)
 
     errors.add :user_authentication, "invalid credentials"
     nil
