@@ -1,5 +1,6 @@
 # frozen_string_literal: true
-require File.expand_path("../boot", __FILE__)
+
+require File.expand_path("boot", __dir__)
 
 require "rails/all"
 
@@ -9,7 +10,7 @@ Bundler.require(*Rails.groups)
 
 module LifeplansApi
   class Application < Rails::Application
-    config.active_record.raise_in_transactional_callbacks = true
     config.autoload_paths << Rails.root.join("lib")
+    config.api_only = true
   end
 end
