@@ -1,6 +1,7 @@
 # frozen_string_literal: true
+
 # Load the Rails application.
-require File.expand_path("../application", __FILE__)
+require File.expand_path("application", __dir__)
 
 # Initialize the Rails application.
 Rails.application.initialize!
