@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,33 +10,31 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20160210231541) do
+ActiveRecord::Schema.define(version: 2016_02_10_231541) do
 
   create_table "bucketlists", force: :cascade do |t|
-    t.string   "name"
+    t.string "name"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.integer  "user_id"
+    t.integer "user_id"
+    t.index ["user_id"], name: "index_bucketlists_on_user_id"
   end
-
-  add_index "bucketlists", ["user_id"], name: "index_bucketlists_on_user_id"
 
   create_table "items", force: :cascade do |t|
-    t.text     "name"
-    t.integer  "bucketlist_id"
-    t.boolean  "done",          default: false
-    t.datetime "created_at",                    null: false
-    t.datetime "updated_at",                    null: false
+    t.text "name"
+    t.integer "bucketlist_id"
+    t.boolean "done", default: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["bucketlist_id"], name: "index_items_on_bucketlist_id"
   end
 
-  add_index "items", ["bucketlist_id"], name: "index_items_on_bucketlist_id"
-
   create_table "users", force: :cascade do |t|
-    t.string   "username"
-    t.datetime "created_at",      null: false
-    t.datetime "updated_at",      null: false
-    t.string   "password_digest"
-    t.string   "auth_token"
+    t.string "username"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.string "password_digest"
+    t.string "auth_token"
   end
 
 end
