@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require "rails_helper"
 
 RSpec.describe "Logging in to an account", type: :request do
@@ -11,10 +12,10 @@ RSpec.describe "Logging in to an account", type: :request do
       "ACCEPT" => "application/vnd.lifeplans-api.v1+json",
       "HTTP_ACCEPT" => "application/vnd.lifeplans-api.v1+json"
     }
-    post "/auth/login", {
+    post "/auth/login", params: {
       username: user.username,
       password: user.password
-    }, headers
+    }, headers: headers
 
     expect(response.content_type).to eq("application/json")
     expect(response).to have_http_status 200
@@ -25,10 +26,11 @@ RSpec.describe "Logging in to an account", type: :request do
       "ACCEPT" => "application/vnd.lifeplans-api.v1+json",
       "HTTP_ACCEPT" => "application/vnd.lifeplans-api.v1+json"
     }
-    post "/auth/login", {
+    post "/auth/login", params: {
       username: user.username,
       password: Faker::Name.name
-    }, headers
+    }, headers: headers
+
     expect(response.content_type).to eq("application/json")
     expect(response).to have_http_status 401
   end
