@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require "rails_helper"
 
 RSpec.describe "Show single bucketlist", type: :request do
@@ -16,9 +17,9 @@ RSpec.describe "Show single bucketlist", type: :request do
       "HTTP_ACCEPT" => "application/vnd.lifeplans-api.v1+json",
       "HTTP_AUTHORIZATION" => "token #{token}"
     }
-    get "/bucketlists/#{bucketlist.id}",
-        {},
-        headers
+
+    get "/bucketlists/#{bucketlist.id}", headers: headers
+
     expect(response.content_type).to eq("application/json")
     expect(response).to have_http_status 200
   end
@@ -33,9 +34,8 @@ RSpec.describe "Show single bucketlist", type: :request do
       "HTTP_ACCEPT" => "application/vnd.lifeplans-api.v1+json",
       "HTTP_AUTHORIZATION" => "token #{token}"
     }
-    get "/bucketlists/#{bucketlist.id + 5}",
-        {},
-        headers
+    get "/bucketlists/#{bucketlist.id + 5}", headers: headers
+
     expect(response.content_type).to eq("application/json")
     expect(response).to have_http_status 404
   end
